Refactor methods - cleaning facade
